Add additional flags to CMakeLists.txt

+++ src/solver.cpp
@@ -16,7 +16,7 @@ struct Context {
     const std::function<void(const vector<Meal> &, float)> &callback_fn;
 };
 
-static void backtracking(const Context &context, int step, vector<Meal> &meals) {
+static void backtracking(const Context &context, unsigned long int step, vector<Meal> &meals) {
     if (step == context.descriptors.size()) {
         float total_cost = std::accumulate(meals.begin(), meals.end(), 0.f, [&](float sum, Meal &meal) {
             Store &store = (Store &) context.store;
